final push

+++ src/main/java/dev/team4/portfoliotracker/services/UserDetailsService.java
@@ -15,18 +15,41 @@ public class UserDetailsService implements org.springframework.security.core.use
     @Autowired
     private UserRepository userRepository;
 
+    /**
+     * Adds user to the database
+     * @param user
+     * @return added User
+     */
     public User createUser(dev.team4.portfoliotracker.models.User user){
         return userRepository.save(user);
     }
 
+    /**
+     * Deletes user matching username from the database
+     * @param user
+     */
     public void removeUser(dev.team4.portfoliotracker.models.User user){
-        userRepository.delete(user);
+        if(user.getUsername() == null){
+            throw new UsernameNotFoundException(user.getUsername());
+        } else {
+            userRepository.delete(user);
+        }
     }
 
+    /**
+     * Gets User matching username
+     * @param username
+     * @return User matching username
+     */
     public User getUserByUsername(String username){
         return userRepository.findByUsername(username);
     }
 
+    /**
+     * stores User matching username in secure UserDetails
+     *@param username
+     *@return UserDetails holding user
+     */
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         User user = userRepository.findByUsername(username);
